moving common code from stackpage to home page
